+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## version 2.0.1: minor feature release
 
 - Introduce `ansible_api.has_var()` method
-- Fix variable transmission in ``
+- Add `defaults` and `overrides` parameters to `run_action()` method; fix `vars` parameter of same in some cases
 - Ability to pass supplemental parameters to `ansible_api.run_action()` through the `Subaction` class
 
 ## version 2.0.0: breaking API change